AC-595, AC-1514: billing api bugfix, refactoring, tests
fixed Kube.is_default unique constraint for "False"
Package prefix and suffix are now limited to 16 symbols

If user have pods with kube type "X", then attempt to remove this kube type from
package will return an error "KubeInUse".

GET /pricing/package/<pkg>/kubes/<kube> returns kube type with price for
this pkg (it had been 405, before)

/usage/<uid>, /users/loginA: if user was not found, returns UserNotFound 404
(before, it had been 400 and without the type)

update (/users/full/<uid>, /editself) returns updated User
(before, it had been just an empty "OK")

added /full/<uid> get full info about one user
(so now /full endpoint have full CRUD)

All known errors are returned with "type":
    UserNotFound
    PackageNotFound
    KubeNotFound
    DuplicateName
    DefaultKubeNotRemovable
    OperationOnInternalKube
    KubeInUse
    PackageInUse

kubedock.utils.atomic decorator will raise an error in case of a commit or
rollback inside atomic block (before commit happens).

+++ kubedock/api/pricing.py
@@ -1,18 +1,49 @@
 from flask import Blueprint, current_app
 from flask.views import MethodView
-from sqlalchemy.exc import IntegrityError, InvalidRequestError
 
 from ..core import db
 from ..rbac import check_permission
 from ..decorators import login_required_or_basic_or_token
-from ..utils import KubeUtils, register_api
+from ..utils import KubeUtils, register_api, atomic
 from ..users import User
 from ..validation import check_pricing_api, package_schema, kube_schema, packagekube_schema
 from ..billing.models import Package, Kube, PackageKube
+from ..pods.models import Pod
 from . import APIError
 
 
 pricing = Blueprint('pricing', __name__, url_prefix='/pricing')
+
+
+class PackageNotFound(APIError):
+    message = 'Package not found'
+    status_code = 404
+
+
+class KubeNotFound(APIError):
+    message = 'Kube not found'
+    status_code = 404
+
+
+class DuplicateName(APIError):
+    pass
+
+
+class DefaultKubeNotRemovable(APIError):
+    pass
+
+
+class OperationOnInternalKube(APIError):
+    status_code = 403
+
+
+class KubeInUse(APIError):
+    pass
+
+
+class PackageInUse(APIError):
+    pass
+
 
 # === PACKAGE ROUTINES ===
 
@@ -21,13 +52,13 @@ pricing = Blueprint('pricing', __name__, url_prefix='/pricing')
 @KubeUtils.jsonwrap
 @login_required_or_basic_or_token
 @check_permission('get', 'pods')
-def get_user_package():
+def get_user_kube_types():
     user = KubeUtils._get_current_user()
     user = User.query.filter_by(username=user.username).first()
     if user is None:
-        raise APIError('No such user', 404)
+        raise APIError('No such user', 404, 'UserNotFound')
     # current_app.logger.debug(user.package.kubes)
-    return {k.kubes.name: k.kubes.id for k in user.package.kubes if k.kubes is not None}
+    return {k.kube.name: k.kube.id for k in user.package.kubes}
 
 
 class PackagesAPI(KubeUtils, MethodView):
@@ -39,60 +70,51 @@ class PackagesAPI(KubeUtils, MethodView):
             return [p.to_dict() for p in Package.query.all()]
         data = Package.query.get(package_id)
         if data is None:
-            raise APIError('Package not found', 404)
+            raise PackageNotFound()
         return data.to_dict()
 
+    @atomic(APIError('Could not create package', 500), nested=False)
     @check_permission('create', 'users')
     def post(self):
         params = check_pricing_api(self._get_params(), package_schema)
         if Package.query.filter_by(name=params['name']).first() is not None:
-            raise APIError('Package with name \'{0}\' already exists'
-                           .format(params['name']))
+            raise DuplicateName('Package with name \'{0}\' already exists'
+                                .format(params['name']))
         package = Package(**params)
         db.session.add(package)
-        try:
-            db.session.commit()
-        except (IntegrityError, InvalidRequestError):
-            db.session.rollback()
-            raise APIError('could not create package')
+        db.session.flush()
         return package.to_dict()
 
+    @atomic(APIError('Could not update package', 500), nested=False)
     @check_permission('edit', 'users')
     def put(self, package_id):
         package = Package.query.get(package_id)
         if package is None:
-            raise APIError('Package not found', 404)
+            raise PackageNotFound()
         params = check_pricing_api(self._get_params(), package_schema, update=True)
 
         if 'name' in params:
             duplicate = Package.query.filter(Package.name == params['name'],
                                              Package.id != package_id).first()
             if duplicate is not None:
-                raise APIError('Package with name \'{0}\' already exists'
-                               .format(params['name']))
+                raise DuplicateName('Package with name \'{0}\' already exists'
+                                    .format(params['name']))
 
         for key, value in params.iteritems():
             setattr(package, key, value)
-        try:
-            db.session.commit()
-        except (IntegrityError, InvalidRequestError):
-            db.session.rollback()
-            raise APIError('could not update package')
+        db.session.flush()
         return package.to_dict()
 
+    @atomic(APIError('Could not delete package', 500), nested=False)
     @check_permission('delete', 'users')
     def delete(self, package_id):
         package = Package.query.get(package_id)
         if package is None:
-            raise APIError('Package not found', 404)
+            raise PackageNotFound()
         if package.users:
-            raise APIError('You have users with this package')
-        try:
-            PackageKube.query.filter_by(package_id=package_id).delete()
-            db.session.delete(package)
-            db.session.commit()
-        except (IntegrityError, InvalidRequestError):
-            raise APIError('could not delete package')
+            raise PackageInUse('You have users with this package')
+        PackageKube.query.filter_by(package_id=package_id).delete()
+        db.session.delete(package)
 
 register_api(pricing, PackagesAPI, 'packages', '/packages/', 'package_id', strict_slashes=False)
 
@@ -109,7 +131,7 @@ class KubesAPI(KubeUtils, MethodView):
             return [i.to_dict() for i in Kube.public_kubes()]
         item = Kube.query.get(kube_id)
         if item is None:
-            raise APIError('Kube not found', 404)
+            raise KubeNotFound()
         return item.to_dict()
 
     @check_permission('create', 'users')
@@ -117,79 +139,68 @@ class KubesAPI(KubeUtils, MethodView):
         params = self._get_params()
         return add_kube(params)
 
+    @atomic(APIError('Could not update kube', 500), nested=False)
     @check_permission('edit', 'users')
     def put(self, kube_id):
         kube = Kube.query.get(kube_id)
         if kube is None:
-            raise APIError('Kube not found', 404)
-        if not Kube.is_kube_editable(kube_id):
-            raise APIError('Kube type is not editable', 403)
+            raise KubeNotFound()
+        if not Kube.is_kube_editable(kube.id):
+            raise OperationOnInternalKube('Kube type is not editable')
         data = check_pricing_api(self._get_params(), kube_schema, update=True)
         if 'name' in data:
             duplicate = Kube.get_by_name(data['name'], Kube.id != kube_id)
             if duplicate is not None:
-                raise APIError('Kube with name \'{0}\' already exists. '
-                               'Name should be unique'.format(data['name']))
+                raise DuplicateName('Kube with name \'{0}\' already exists. '
+                                    'Name should be unique'.format(data['name']))
         is_default = data.get('is_default', None)
         if is_default:
             _remove_is_default_kube_flags()
         elif kube.is_default and is_default is not None:
-            # Do not allow to turn off is_default flag from a kube, because in
-            # this case there will be no default kube. The only right (and
-            # acceptable) way to remove this flag: set is_default = True for
-            # another kube.
-            raise APIError("Can't remove default flag for the kube")
+            raise DefaultKubeNotRemovable(
+                'Setting "is_default" flag to false is forbidden. You can change '
+                'default kube type by setting another kube type as default.')
 
         for key, value in data.items():
             setattr(kube, key, value)
-        try:
-            db.session.commit()
-        except (IntegrityError, InvalidRequestError):
-            db.session.rollback()
-            raise APIError('could not update kube')
+        db.session.flush()
         return kube.to_dict()
 
+    @atomic(APIError('Could not delete kube', 500), nested=False)
     @check_permission('delete', 'users')
     def delete(self, kube_id):
         kube = Kube.query.get(kube_id)
         if kube is None:
-            raise APIError('Kube not found', 404)
+            raise KubeNotFound()
         if kube.is_default:
-            raise APIError(
+            raise DefaultKubeNotRemovable(
                 'Deleting of default kube type is forbidden. '
-                'Set another kube type as default and try again', 403)
-        if not Kube.is_kube_editable(kube_id):
-            raise APIError('Kube type is not editable', 403)
-        if kube.nodes:
-            raise APIError('Some nodes use this kube type')
+                'Set another kube type as default and try again')
+        if not Kube.is_kube_editable(kube.id):
+            raise OperationOnInternalKube('Kube type is not editable')
         if kube.pods:
-            raise APIError('Some pods use this kube type')
-        try:
-            PackageKube.query.filter_by(kube_id=kube_id).delete()
-            db.session.delete(kube)
-            db.session.commit()
-        except (IntegrityError, InvalidRequestError):
-            raise APIError('could not delete package')
+            raise KubeInUse('Some pods use this kube type')
+        if kube.nodes:
+            raise KubeInUse('Some nodes use this kube type')
+        PackageKube.query.filter_by(kube_id=kube_id).delete()
+        db.session.delete(kube)
 
 register_api(pricing, KubesAPI, 'kubes', '/kubes/', 'kube_id', strict_slashes=False)
 
 
+@atomic(APIError('Could not create kube', 500), nested=False)
 def add_kube(data):
     data = check_pricing_api(data, kube_schema)
     if Kube.get_by_name(data['name']) is not None:
-        raise APIError('Kube type with name \'{0}\' already exists. '
-                       'Name should be unique'.format(data['name']))
+        raise DuplicateName('Kube type with name \'{0}\' already exists. '
+                            'Name must be unique'.format(data['name']))
 
     kube = Kube(**data)
     if kube.is_default:
         # reset is_default flag for all kubes if new kube marked as default
         _remove_is_default_kube_flags()
     db.session.add(kube)
-    try:
-        db.session.commit()
-    except (IntegrityError, InvalidRequestError):
-        db.session.rollback()
-        raise APIError('could not create package')
+    db.session.flush()
     return kube.to_dict()
 
 
@@ -204,8 +215,8 @@ def add_kube(data):
 def get_package_kube_ids(package_id):
     package = Package.query.get(package_id)
     if package is None:
-        raise APIError('Package not found', 404)
-    return [kube.kube_id for kube in package.kubes if kube.kubes is not None]
+        raise PackageNotFound()
+    return [kube.kube_id for kube in package.kubes]
 
 
 @pricing.route('/packages/<int:package_id>/kubes-by-name', methods=['GET'],
@@ -216,8 +227,8 @@ def get_package_kube_ids(package_id):
 def get_package_kube_names(package_id):
     package = Package.query.get(package_id)
     if package is None:
-        raise APIError('Package not found', 404)
-    return [kube.kubes.name for kube in package.kubes if kube.kubes is not None]
+        raise PackageNotFound()
+    return [package_kube.kube.name for package_kube in package.kubes]
 
 
 class PackageKubesAPI(KubeUtils, MethodView):
@@ -225,61 +236,67 @@ class PackageKubesAPI(KubeUtils, MethodView):
 
     @check_permission('get', 'users')
     def get(self, package_id, kube_id=None):
-        if kube_id is not None:
-            raise APIError('Method not allowed', 405)
         package = Package.query.get(package_id)
         if package is None:
-            raise APIError('Package not found', 404)
-        return [dict(kube.kubes.to_dict(), kube_price=kube.kube_price)
-                for kube in package.kubes if kube.kubes is not None]
+            raise PackageNotFound()
+        if kube_id is None:
+            return [dict(pk.kube.to_dict(), kube_price=pk.kube_price)
+                    for pk in package.kubes]
+        try:
+            return (dict(pk.kube.to_dict(), kube_price=pk.kube_price)
+                    for pk in package.kubes if pk.kube_id == int(kube_id)).next()
+        except StopIteration:
+            raise KubeNotFound()
 
+    @atomic(APIError('Could not add kube type to package', 500), nested=False)
     @check_permission('create', 'users')
     def post(self, package_id):
         if Package.query.get(package_id) is None:
-            raise APIError('Package not found', 404)
+            raise PackageNotFound()
         params = self._get_params()
         if 'id' in params:
             params = check_pricing_api(params, packagekube_schema)
             kube_id = params['id']
             kube = Kube.query.get(kube_id)
             if kube is None:
-                raise APIError('Kube not found', 404)
+                raise KubeNotFound()
             if not kube.is_public():
-                raise APIError(
-                    'Kube type is not allowed to use it in packages', 403)
+                raise OperationOnInternalKube(
+                    'Kube type is not allowed to use it in packages')
         else:
             params = check_pricing_api(params, dict(kube_schema, **packagekube_schema))
             kube_id = add_kube({key: value for key, value in params.iteritems()
-                                if key in kube_schema})['id']
+                                if key in kube_schema}, commit=False)['id']
 
         return _add_kube_type_to_package(package_id, kube_id, params['kube_price'])
 
+    @atomic(APIError('Could not update kube type in package', 500), nested=False)
     @check_permission('edit', 'users')
     def put(self, package_id=None, kube_id=None):
         if Package.query.get(package_id) is None:
-            raise APIError('Package not found', 404)
+            raise PackageNotFound()
         kube = Kube.query.get(kube_id)
         if kube is None:
-            raise APIError('Kube not found', 404)
+            raise KubeNotFound()
         if not kube.is_public():
-            raise APIError(
-                'Kube type is not allowed to use it in packages', 403)
+            raise OperationOnInternalKube(
+                'Kube type is not allowed to use it in packages')
         params = check_pricing_api(self._get_params(), packagekube_schema)
 
         return _add_kube_type_to_package(package_id, kube_id, params['kube_price'])
 
+    @atomic(APIError('Could not remove kube type from package', 500), nested=False)
     @check_permission('delete', 'users')
     def delete(self, package_id, kube_id):
         package_kube = PackageKube.query.filter_by(package_id=package_id,
                                                    kube_id=kube_id).first()
         if package_kube is None:
-            raise APIError('Kube type is not in the package', 404)
+            raise KubeNotFound('Kube type is not in the package')
+        if User.query.filter(User.pods.any(Pod.kube == package_kube.kube),
+                             User.package == package_kube.package).first():
+            raise KubeInUse('Some users with this package have pods '
+                            'with this kube type')
         db.session.delete(package_kube)
-        try:
-            db.session.commit()
-        except (IntegrityError, InvalidRequestError):
-            db.session.rollback()
-            raise APIError('could not remove kube type from package')
 
 register_api(pricing, PackageKubesAPI, 'packagekubes',
              '/packages/<int:package_id>/kubes/', 'kube_id', strict_slashes=False)
@@ -291,16 +308,9 @@ def _add_kube_type_to_package(package_id, kube_id, kube_price):
         package_kube = PackageKube(package_id=package_id, kube_id=kube_id)
         db.session.add(package_kube)
     package_kube.kube_price = kube_price
-
-    try:
-        db.session.commit()
-    except (IntegrityError, InvalidRequestError):
-        db.session.rollback()
-        raise APIError('could not add kube type to package')
+    db.session.flush()
     return package_kube.to_dict()
 
 
 def _remove_is_default_kube_flags():
-    db.session.query(Kube).update(
-        {Kube.is_default: None}, synchronize_session='fetch'
-    )
+    Kube.query.update({Kube.is_default: None}, synchronize_session='fetch')
